Fixed handling of relative paths. General cleanup.
The code could not handle relative paths since the working
directory is changed to run PAML. Solution was to make sure all paths
were made absolute (this is not necessarily required) and to change the
working directory back after each run of PAML to the starting directory.

This could be simplified by forcing the passing of full absolute paths
by the user.